requirements.txt
Flask==2.2.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
PyJWT~=2.9.0
pymongo==4.3.3
boto3==1.26.36
firebase-admin==6.0.0
bcrypt==4.0.1
cryptography>=3.4.7
gunicorn==20.1.0
Werkzeug==2.2.2
